memory calculated per model

+++ classify.py
@@ -215,7 +215,7 @@ def convert_labels(labels):
 def load_sparse_csr(filename):
     loader = np.load(filename)
     return csr_matrix((loader['data'], loader['indices'], loader['indptr']),
-                         shape=loader['shape'], dtype="int16"), loader['labels']
+                         shape=loader['shape'], dtype="float32"), loader['labels']
 
 
 def load_data(size, file2, file3):
@@ -292,6 +292,7 @@ def main(size='sm', file2='0', file3='0', red='0', tfidf='0', prune='0', est='32
         tfer = TfidfTransformer()
         tfer.fit(features)
         features = tfer.transform(features)
+        features = features.astype('float32')
 
     print "Final data shape:", features.shape
     logging.info("Final data shape: %s" % (features.shape,))
@@ -309,7 +310,7 @@ def main(size='sm', file2='0', file3='0', red='0', tfidf='0', prune='0', est='32
         print "Time elapsed for dimensionality reduction is %f" %  elapsed
         logging.info("Time elapsed for dimensionality reduction is %f" %  elapsed)
 
-    features = features.astype('float32')
+    #features = features.astype('float32')
 
     results = classify_all(labels, features, clfs, folds, model_names)
     #results.sort("Split Val Acc", inplace=True, ascending=False)
